Group repository classes into a separate package
