+++ user_management_service/app/storage/repos/PasswordInfoRepository.scala
@@ -1,4 +1,4 @@
-package storage
+package storage.repos
 
 import storage.db.SilhouetteTableRepository
