+++ user_management_service/app/storage/repos/UserRepository.scala
@@ -1,4 +1,4 @@
-package storage
+package storage.repos
 
 import exceptions.StorageException._
 import storage.model.UserResource
